Simple 6-wire character LCD interface objects
Objects to print strings on character LCDs (44780-type), connected
via simple interface, with 4 bits of data (D4..D7), E and RS,
and assuming R/W is always tied low (i.e. write). This is the
way the display is connected in the Audiothingies P6 which is
what these objects have been tested on.

lcdctrl.axo is the LCD controller object, which handles the hardware
interfacing and thus must be set up accordingly.
lcdprint.axo is the actual printing object, which refers to a controller
object and has a trigger input for printing the required string
to the LCD.