Objects for reading from 74HC165 shift registers
In many designs, 74HC165 shift registers are used to read digital
inputs using an SPI type of interface. These two objects implement
inputs connected 74HC165 shift registers.

There is a control object which contains the port assignemnts for
the data, clock, latch and inhibit pins on the 74HC165 chips.
The bits attribute indicates how many 74HC165s are connected; there
are 8 bits with one 74HC165, 16 bits with two, etc.
Finally, the rate attribute controls how fast the input chain is
scanned. The fast the input is expected to change (e.g. if
it is an encoder), the higher the rate needs to be (most easily
determined empirically).

For each bit connected to the 74HC165(s), a 165in object can be
used to extract the requred bit, as set in the corresponding
attribute. The port attribute must refer to the (name of the)
corresponding 165ctrl object.