Objects to control outputs connected to 74HC595 shift registers
In many designs, 74HC595 shift registers are used to provide
output bits using a serial protocol like SPI. These objects
provide access to individual bits connected to shift
registers, either on a single 75HC595 or multiple ones connected
in series.

The control object 595ctrl contains attributes for the pin connections
for the data, clock, latch and output enable pins on the 74HC595s,
as well as the number of bits available - 8 for one 74HC595, 16
for two, etc.

Access to individual bits is then achieved using the 595out object,
which refers to a specific 595ctrl port and a bit number.