Quadrature decoder object
The qudecode.axo object is a quadrature decoder for decoding
rotary encoders connected to two digital inputs. The inputs
on the object are designated a and b, than there are then three
clock outputs: one which outputs clocks when the encoder is
turned in one direction, and one when it is turned in the other
direction, and one (step) when it is turned in either direction.

The rate attribute determines how many clocks are output per
input state change. When the rate is set to full, one clock is
output per input state change on any one of the a or b inputs.
When the rate is set to half, a clock is output only when there
is a state change on the a input. When set to quarter, a clock
is output for positive going state changes on the a input only.
The rate is used to determine how fast the outputs are clocked
in relation to the knob speed.